requirements.txt
python-dotenv==1.0.0
requests
pandas
openai
instructor
pydantic
gspread
oauth2client
gspread-dataframe